move db fn elsewhere
