+++ docker-compose.yml
@@ -6,6 +6,6 @@ services:
     ports:
       - "5433:5432"
     environment:
-      POSTGRES_PASSWORD: notsecurepassword
+      - POSTGRES_PASSWORD=notsecurepassword
     volumes:
       - ./init.sql:/docker-entrypoint-initdb.d/init.sql
